+++ CHANGELOG.md
@@ -23,6 +23,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added compatibility if a Windows FTP server returns file information to the
   `LIST` command with 24-hour times. Closes [#438](https://github.com/PyFilesystem/pyfilesystem2/issues/438).
 
+### Added
+
+- Added `passwd` argument and `setpassword` for ReadZipFS to extract password
+  protected date from zip file. [#360](https://github.com/PyFilesystem/pyfilesystem2/issues/360)
+
 ### Changed
 
 - Start testing on PyPy. Due to [#342](https://github.com/PyFilesystem/pyfilesystem2/issues/342)